Fix compilation errors in examples and library
✅ Fixed compilation issues:
- Fixed width() method calls (use f32 instead of integer)
- Removed height() method calls (not available on PlotSeriesBuilder)
- Added .build() calls before .save() for proper Plot construction
- Fixed type annotations in zero_copy.rs test

📊 Examples now compile properly:
- examples/basic_interaction.rs
- examples/real_time_performance.rs
- examples/create_performance_plot.rs
- examples/gpu_scaling_benchmark.rs

🛠️ Library compilation:
- Core library compiles with warnings only
- Interactive features disabled for compatibility
- Zero unsafe code maintained in public API

+++ src/data/zero_copy.rs
@@ -370,8 +370,8 @@ mod tests {
         assert_eq!(mapped.get(4), Some(10.0));
         
         // Test Data1D implementation on mapped view
-        assert_eq!(mapped.get(0).unwrap().into(), 2.0);
-        assert_eq!(mapped.get(4).unwrap().into(), 10.0);
+        assert_eq!(mapped.get(0).unwrap().into() as f64, 2.0);
+        assert_eq!(mapped.get(4).unwrap().into() as f64, 10.0);
     }
 
     #[test]
